==> AI-Executive-Decision-Platform/notebooks/01_data_generation.ipynb <==
import numpy as np
print(np.__version__)


import pandas as pd
import numpy as np
from datetime import datetime, timedelta

np.random.seed(42)

n_customers = 10000

customers = pd.DataFrame({
    "customer_id": range(1, n_customers + 1),
    "region": np.random.choice(["North", "South", "East", "West"], n_customers),
    "age": np.random.randint(18, 60, n_customers),
    "purchase_frequency": np.random.poisson(5, n_customers),
    "avg_order_value": np.random.normal(2000, 500, n_customers),
    "discount_usage_rate": np.random.uniform(0, 1, n_customers),
    "last_purchase_days": np.random.randint(1, 365, n_customers)
})

customers["total_spend"] = customers["purchase_frequency"] * customers["avg_order_value"]

# Simple churn logic
customers["churn_flag"] = np.where(
    customers["last_purchase_days"] > 200,
    1,
    0
)

customers.to_csv("../data/customers.csv", index=False)

print("Customers dataset created ✅")

df=pd.read_csv('../data/customers.csv')
df

n_transactions = 50000

start_date = datetime(2023, 1, 1)

transactions = pd.DataFrame({
    "transaction_id": range(1, n_transactions + 1),
    "customer_id": np.random.choice(customers["customer_id"], n_transactions),
    "product_category": np.random.choice(["Electronics", "Fashion", "Home", "Beauty"], n_transactions),
    "price": np.random.normal(2500, 800, n_transactions),
    "discount": np.random.uniform(0, 0.3, n_transactions),
    "marketing_spend": np.random.uniform(50, 300, n_transactions),
    "quantity": np.random.randint(1, 4, n_transactions),
    "date": [start_date + timedelta(days=int(x)) for x in np.random.randint(0, 730, n_transactions)]
})

transactions["cost"] = transactions["price"] * 0.6
transactions["revenue"] = transactions["price"] * transactions["quantity"] * (1 - transactions["discount"])
transactions["profit"] = transactions["revenue"] - (transactions["cost"] * transactions["quantity"]) - transactions["marketing_spend"]

transactions.to_csv("../data/transactions.csv", index=False)
print("Transactions dataset created ✅")

transactions["month"] = transactions["date"].dt.to_period("M")

monthly_summary = transactions.groupby("month").agg({
    "revenue": "sum",
    "profit": "sum",
    "marketing_spend": "sum",
    "quantity": "sum"
}).reset_index()

monthly_summary["month"] = monthly_summary["month"].astype(str)

monthly_summary.to_csv("../data/monthly_summary.csv", index=False)

print("Monthly summary created ✅")
